more casts

+++ kernels/matmul_transb/main.c
@@ -18,9 +18,9 @@ int main() {
 
     // Copy data in shared local memory
     if (snrt_is_dm_core()) {
-        snrt_dma_start_1d(local_x, X, M * K * sizeof(DTYPE));
-        snrt_dma_start_1d(local_y, Y, N * K * sizeof(DTYPE));
-        snrt_dma_start_1d(local_z, G_IN, M * N * sizeof(DTYPE));
+        snrt_dma_start_1d((uint64_t)local_x, (uint64_t)X, M * K * sizeof(DTYPE));
+        snrt_dma_start_1d((uint64_t)local_y, (uint64_t)Y, N * K * sizeof(DTYPE));
+        snrt_dma_start_1d((uint64_t)local_z, (uint64_t)G_IN, M * N * sizeof(DTYPE));
         snrt_dma_wait_all();
     }
